Minor refactor some some in vehicles.
