+++ Resources/Textures/_Starlight/Objects/Devices/pda.rsi/meta.json
@@ -62,9 +62,6 @@
     {
       "name": "pda-veteranmining"
     },
-	  {
-      "name": "pda-seniorchemist"
-    },
     {
       "name": "pda-veteranbrigmedic"
     },
